requirements.txt
bleach==3.1.3
Django==2.2
django-ckeditor==6.2.0
django-js-asset==1.2.2
Pillow==8.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
